+++ src/main/java/de/bauersoft/views/course/CourseView.java
@@ -1,13 +1,18 @@
 package de.bauersoft.views.course;
 
+import com.vaadin.flow.component.Text;
+import com.vaadin.flow.component.grid.GridVariant;
 import com.vaadin.flow.component.grid.contextmenu.GridContextMenu;
+import com.vaadin.flow.component.grid.contextmenu.GridMenuItem;
 import com.vaadin.flow.component.html.Div;
+import com.vaadin.flow.component.notification.Notification;
+import com.vaadin.flow.component.notification.NotificationVariant;
 import com.vaadin.flow.router.PageTitle;
 import com.vaadin.flow.router.Route;
-
 import de.bauersoft.components.autofiltergrid.AutoFilterGrid;
-import de.bauersoft.data.entities.Course;
+import de.bauersoft.data.entities.course.Course;
 import de.bauersoft.data.providers.CourseDataProvider;
+import de.bauersoft.services.ComponentService;
 import de.bauersoft.services.CourseService;
 import de.bauersoft.views.DialogState;
 import de.bauersoft.views.MainLayout;
@@ -16,22 +21,67 @@ import jakarta.annotation.security.RolesAllowed;
 @PageTitle("Course")
 @Route(value = "course", layout = MainLayout.class)
 @RolesAllowed(value = { "ADMIN", "ACCOUNTENT" })
-public class CourseView extends Div {
-	
+public class CourseView extends Div
+{
 	AutoFilterGrid<Course> grid = new AutoFilterGrid<Course>(Course.class, false, true);
 
-	public CourseView(CourseService service,  CourseDataProvider dataProvider) {
+	public CourseView(CourseService service,
+					  CourseDataProvider dataProvider,
+					  ComponentService componentService)
+	{
 		setClassName("content");
-		grid.addColumn("name");
-		grid.addItemDoubleClickListener(event -> new CourseDialog(service,dataProvider, event.getItem(), DialogState.EDIT));
+
 		grid.setDataProvider(dataProvider);
-		GridContextMenu<Course> contextMenu = grid.addContextMenu();
-		contextMenu.addItem("new", event -> new CourseDialog(service,dataProvider, new Course(), DialogState.NEW));
-		contextMenu.addItem("delete", event -> event.getItem().ifPresent(item -> {
-			service.delete(item.getId());
-			dataProvider.refreshAll();
-		}));
+
 		grid.setHeightFull();
+		grid.setWidthFull();
+		grid.addThemeVariants(GridVariant.LUMO_ROW_STRIPES);
+
+		grid.addColumn("name");
+
+		grid.addItemDoubleClickListener(event ->
+		{
+			new CourseDialog(service,dataProvider, event.getItem(), DialogState.EDIT);
+		});
+
+		GridContextMenu<Course> contextMenu = grid.addContextMenu();
+		contextMenu.addItem("new", event ->
+		{
+			new CourseDialog(service,dataProvider, new Course(), DialogState.NEW);
+		});
+
+		GridMenuItem<Course> deleteItem = contextMenu.addItem("delete", event ->
+		{
+			event.getItem().ifPresent(item ->
+			{
+				if(componentService.getRepository().existsByCourseId(item.getId()))
+				{
+					Div div = new Div();
+					div.setMaxWidth("33vw");
+					div.getStyle().set("white-space", "normal");
+					div.getStyle().set("word-wrap", "break-word");
+
+					div.add(new Text("Der Course \"" + item.getName() + "\" kann nicht gelöscht werden da er noch in einigen Components verwendet wird."));
+
+					Notification notification = new Notification(div);
+					notification.setDuration(5000);
+					notification.setPosition(Notification.Position.MIDDLE);
+					notification.addThemeVariants(NotificationVariant.LUMO_WARNING);
+					notification.open();
+
+					return;
+				}
+
+				service.delete(item.getId());
+				dataProvider.refreshAll();
+			});
+		});
+
+		contextMenu.addGridContextMenuOpenedListener(event ->
+		{
+			deleteItem.setVisible(event.getItem().isPresent());
+		});
+
 		this.add(grid);
 	}
 }
